Measure memory in fractional kilobytes rather than rounding down.
git-svn-id: http://ruby-prof.rubyforge.org/svn@260 5fb4a865-4516-0410-ae3b-a609ddf12040

+++ ext/measure_memory.h
@@ -36,7 +36,7 @@ measure_memory()
 static double
 convert_memory(prof_measure_t c)
 { 
-    return  c / 1024; 
+    return (double) c / 1024; 
 }
 
 #endif
